chore: improve UX of `forc index` plugin commands (#1142)
* Remove forc index new

* Allow for optional namespace in forc index init

* Remove profile and target flags from forc index build; change release bool to debug

* Remove service check from forc index check

* Move service check into forc index status

* Cleanup

* Remove short arg collision from forc index postgres

* Fix CI

* Replace forc_index_init with forc_index_new

* Adjust tests and documentation

* Address feedback; clean up some other things

* Address everything for real this time
