+++ .github/workflows/ci.yml
@@ -305,14 +305,14 @@ jobs:
 
           # WASM indexer
           forc-index new indexer-test --namespace fuel
-          cd indexer-test/
+          cd indexer-test
           forc-index build
           cd ..
           rm -rfv indexer-test
 
           # Native indexer
           forc-index new indexer-test --namespace fuel --native
-          cd indexer-test/
+          cd indexer-test
           forc-index build --native
           cd ..
           rm -rfv indexer-test
