--- docs/src/forc-index/init.md
@@ -1,23 +1,0 @@
-# `forc index init`
-
-Create a new indexer project in the current directory.
-
-```bash
-forc index init --namespace fuel
-```
-
-```text
-Create a new indexer project in the current directory
-
-USAGE:
-    forc-index init [OPTIONS] --namespace <NAMESPACE>
-
-OPTIONS:
-        --absolute-paths           Resolve indexer asset filepaths using absolute paths.
-    -h, --help                     Print help information
-        --name <NAME>              Name of indexer.
-        --namespace <NAMESPACE>    Namespace to which indexer belongs.
-        --native                   Initialize an indexer with native execution enabled.
-    -p, --path <PATH>              Path at which to create indexer.
-    -v, --verbose                  Enable verbose output.
-```
