+++ docs/src/forc-index/new.md
@@ -3,12 +3,12 @@
 Create a new indexer project in a new directory.
 
 ```bash
-forc index new --namespace fuel --path /home/fuel/projects
+forc index new --namespace fuel /home/fuel/projects/example_indexer
 ```
 
 ```text
 USAGE:
-    forc-index new [OPTIONS] --namespace <NAMESPACE> <PATH>
+    forc-index new [OPTIONS] <PATH>
 
 ARGS:
     <PATH>    Path at which to create indexer
@@ -18,7 +18,6 @@ OPTIONS:
     -h, --help                     Print help information
         --name <NAME>              Name of indexer.
         --namespace <NAMESPACE>    Namespace to which indexer belongs.
-        --native                   Whether to initialize an indexer with native execution enabled.
-    -v, --verbose <verbose>        Enable verbose output. [default: true]
-
+        --native                   Initialize an indexer with native execution enabled.
+    -v, --verbose                  Enable verbose output
 ```
