+++ plugins/forc-index/src/commands/mod.rs
@@ -2,7 +2,6 @@ pub mod auth;
 pub mod build;
 pub mod check;
 pub mod deploy;
-pub mod init;
 pub mod kill;
 pub mod new;
 pub mod remove;
