+++ plugins/forc-index/src/commands/new.rs
@@ -6,22 +6,19 @@ use std::path::PathBuf;
 /// Create a new indexer project in a new directory.
 #[derive(Debug, Parser)]
 pub struct Command {
-    /// Name of indexer
+    /// Path at which to create indexer.
+    pub path: PathBuf,
+
+    /// Name of indexer.
     #[clap(long, help = "Name of indexer.")]
     pub name: Option<String>,
 
-    /// Path at which to create indexer
-    pub path: PathBuf,
-
     /// Namespace to which indexer belongs.
     #[clap(long, help = "Namespace to which indexer belongs.")]
-    pub namespace: String,
+    pub namespace: Option<String>,
 
-    /// Whether to initialize an indexer with native execution enabled
-    #[clap(
-        long,
-        help = "Whether to initialize an indexer with native execution enabled."
-    )]
+    /// Initialize an indexer with native execution enabled.
+    #[clap(long, help = "Initialize an indexer with native execution enabled.")]
     pub native: bool,
 
     /// Resolve indexer asset filepaths using absolute paths.
