+++ plugins/forc-index/src/defaults.rs
@@ -10,8 +10,6 @@ pub const CARGO_CONFIG_FILENAME: &str = "config";
 pub const INDEXER_SERVICE_HOST: &str = "http://127.0.0.1:29987";
 pub const WEB_API_PORT: &str = defaults::WEB_API_PORT;
 pub const WASM_TARGET: &str = "wasm32-unknown-unknown";
-pub const INDEXER_TARGET: &str = "wasm32-unknown-unknown";
-pub const BUILD_RELEASE_PROFILE: &str = "true";
 pub const MESSAGE_PADDING: usize = 55;
 pub const SUCCESS_EMOJI_PADDING: usize = 3;
 pub const FAIL_EMOJI_PADDING: usize = 6;
