+++ plugins/forc-index/src/ops/forc_index_check.rs
@@ -1,59 +1,7 @@
-use crate::cli::CheckCommand;
 use crate::defaults;
 use crate::utils::{center_align, find_executable_with_msg, rightpad_whitespace};
-use serde_json::{to_string_pretty, value::Value, Map};
-use std::process::Command;
-use tracing::{error, info};
 
-fn find_indexer_service_info(grpahql_api_port: &str) -> (String, String) {
-    let (emoji, msg) = match Command::new("lsof")
-        .arg(&format!("-ti:{grpahql_api_port}"))
-        .output()
-    {
-        Ok(o) => {
-            let (emoji, msg) = match String::from_utf8_lossy(&o.stdout)
-                .to_string()
-                .strip_suffix('\n')
-            {
-                Some(pid) => (
-                    center_align("✅", defaults::SUCCESS_EMOJI_PADDING),
-                    rightpad_whitespace(
-                        &format!(
-                            "Local service found: PID({pid}) | Port({grpahql_api_port})."
-                        ),
-                        defaults::MESSAGE_PADDING,
-                    ),
-                ),
-                None => (
-                    center_align("⛔️", defaults::FAIL_EMOJI_PADDING),
-                    rightpad_whitespace(
-                        &format!("Failed to detect service at Port({grpahql_api_port}).",),
-                        defaults::MESSAGE_PADDING,
-                    ),
-                ),
-            };
-
-            (emoji, msg)
-        }
-        Err(_e) => (
-            center_align("⛔️", defaults::FAIL_EMOJI_PADDING),
-            rightpad_whitespace(
-                &format!("Failed to detect service at Port({grpahql_api_port}).",),
-                defaults::MESSAGE_PADDING,
-            ),
-        ),
-    };
-
-    (emoji, msg)
-}
-
-pub async fn init(command: CheckCommand) -> anyhow::Result<()> {
-    let CheckCommand {
-        url,
-        graphql_api_port,
-    } = command;
-
-    let target = format!("{url}/api/health");
+pub async fn init() -> anyhow::Result<()> {
     let psql = "psql";
     let fuel_indexer = "fuel-indexer";
     let fuel_core = "fuel-core";
@@ -64,37 +12,11 @@ pub async fn init(command: CheckCommand) -> anyhow::Result<()> {
     let rustc = "rustc";
     let forc_wallet = "forc-wallet";
 
-    match reqwest::get(&target).await {
-        Ok(res) => {
-            if res.status() != reqwest::StatusCode::OK {
-                error!(
-                    "\n❌ {target} returned a non-200 response code: {:?}",
-                    res.status()
-                );
-                return Ok(());
-            }
-
-            let res_json = res
-                .json::<Map<String, Value>>()
-                .await
-                .expect("Failed to read JSON response.");
-
-            info!(
-                "\n✅ Sucessfully fetched service health:\n\n{}",
-                to_string_pretty(&res_json).unwrap()
-            );
-        }
-        Err(e) => {
-            error!("\n❌ Could not connect to indexer service:\n'{e}'");
-        }
-    }
-
     let (indexer_emoji, _indexer_path, indexer_msg) =
         find_executable_with_msg(fuel_indexer);
     let (psql_emoji, _psql_path, psql_msg) = find_executable_with_msg(psql);
     let (fuel_core_emoji, _fuelcore_path, fuel_core_msg) =
         find_executable_with_msg(fuel_core);
-    let (service_emoji, service_msg) = find_indexer_service_info(&graphql_api_port);
     let (docker_emoji, _docker_path, docker_msg) = find_executable_with_msg(docker);
     let (fuelup_emoji, _fuelup_path, fuelup_msg) = find_executable_with_msg(fuelup);
     let (forc_pg_emoji, _forc_pg_path, forc_pg_msg) = find_executable_with_msg(forc_pg);
@@ -111,8 +33,6 @@ pub async fn init(command: CheckCommand) -> anyhow::Result<()> {
     let status_headers = center_align("Status", status_padding);
     let binary_header =
         rightpad_whitespace("fuel-indexer binary", defaults::HEADER_PADDING);
-    let service_header =
-        rightpad_whitespace("fuel-indexer service", defaults::HEADER_PADDING);
     let psql_header = rightpad_whitespace(psql, defaults::HEADER_PADDING);
     let fuel_core_header = rightpad_whitespace(fuel_core, defaults::HEADER_PADDING);
     let docker_header = rightpad_whitespace(docker, defaults::HEADER_PADDING);
@@ -128,8 +48,6 @@ pub async fn init(command: CheckCommand) -> anyhow::Result<()> {
 | {status_headers} |  {check_header}  |{details_header}|
 +--------+------------------------+---------------------------------------------------------+
 |  {indexer_emoji}  | {binary_header}   |  {indexer_msg}|
-+--------+------------------------+---------------------------------------------------------+
-| {service_emoji} | {service_header}   |  {service_msg}|
 +--------+------------------------+---------------------------------------------------------+
 |  {psql_emoji}  | {psql_header}   |  {psql_msg}|
 +--------+------------------------+---------------------------------------------------------+
