+++ plugins/forc-index/src/ops/forc_index_deploy.rs
@@ -23,9 +23,7 @@ pub async fn init(command: DeployCommand) -> anyhow::Result<()> {
         manifest,
         path,
         auth,
-        target,
-        release,
-        profile,
+        debug,
         locked,
         native,
         target_dir,
@@ -39,9 +37,7 @@ pub async fn init(command: DeployCommand) -> anyhow::Result<()> {
         build::exec(BuildCommand {
             manifest: manifest.clone(),
             path: path.clone(),
-            target,
-            release,
-            profile,
+            debug,
             verbose,
             locked,
             native,
