+++ plugins/forc-index/src/ops/mod.rs
@@ -2,7 +2,6 @@ pub mod forc_index_auth;
 pub mod forc_index_build;
 pub mod forc_index_check;
 pub mod forc_index_deploy;
-pub mod forc_index_init;
 pub mod forc_index_kill;
 pub mod forc_index_new;
 pub mod forc_index_remove;
